requirements.txt
# platform: osx-64
python=3.6.13
eth-brownie=1.14.4
pytest=6.0.1
python-dotenv=0.16.0

